refactor(e2e): keep Podman proof linear
